Now generating the tree inside the component for proper title rendering
